Unit Test recovering & writing new ones
Tests:
- Some tests have been recovered;
- GeoChatViewModelTest has been expanded;

Other:
- Some other stuff has been provided as always;

+++ client/app/src/test/java/com/qubacy/geoqq/ui/screen/geochat/settings/GeoChatSettingsViewModelTest.kt
@@ -19,7 +19,7 @@ class GeoChatSettingsViewModelTest : ViewModelTest() {
 
     data class GetCurRadiusOptionMetersTestCase(
         val radiusOptionIndex: Int,
-        val expectedRadiusOptionMeters: Float
+        val expectedRadiusOptionMeters: Int
     )
 
     @Test
